chore: refactor notification card
